move types into separate module
